# swagger.json
{
	"info": {
		"description":"CS 4783 Assignment"
		,"title":"CS 4783","version":"1.0.0" 
	}
	,"host":"localhost:8080"
	
	,"basePath":"/"
	,"produces": [
		"application/json" 
	]
	,"schemes": [
		"https"
	]
	,"securityDefinitions": {
	}
	,"swagger":"2.0"
	,"paths": { 
		"/hello": { 
			"get": {
				"parameters":[ 
				] 
				,"description":"The proverbial hello world minus the world." 
				,"tags":[
					"hello"
				]
				,"responses":{
					"200": {
						"description": "Hello Mundo" 
					}
				}
			}
		}

		,"/properties": {
			"get": {
				"description": "Returns a list of stored US properties ordered by id"
				,"tags": [
					"Get"
				]
				,"responses": {
					"200": {
						"description": "Success"
						,"schema": {
							"$ref": "#/definitions/Property"
						}	
					}	
				}
			},

			"post": {
				"parameters": [
					{
						
						"in":"body",
						"schema": {"$ref": "#/definitions/Property"},
						"required": true,
						"description": "Property object"
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": true,
						"type": "string"
					}
				],
				"description": "Adds a new property"
				,"tags": [
					"Post"
				]
				,"responses": {
					"200": {
						"description": "Success"
					}
				}
			}
		}

		,"/properties/{id}": {
			"parameters": [
				{"name":"id",
					"in":"path",
					"description":"id of property to return",
					"required":true,
					"type":"integer",
					"format":"int64"
				}
			],

			"get": {
				"description": "Returns property if id is found"
				,"tags": [
					"Get by ID"
				]
				,"responses": {
					"200": {
						"description": "Success"
						,"schema": {
							"$ref": "#/definitions/Property"
						}
					}
					,"400": {
						"description": "Bad Request"
					}
					,"404": {
						"description": "Not Found"
					}
				}
			},

			"delete": {
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": true,
						"type": "string"
					}
				],
				"description": "deletes property"
				,"tags": [
					"Delete by ID"
				]
				,"responses": {
					"200": {
						"description": "Deleted"
					}
				}
			},

			"put": {
				"parameters": [
					{
						"in":"body",
						"schema": {"$ref": "#/definitions/Property"},
						"required": true,
						"description": "Property object"
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": true,
						"type": "string"
					}
				],
				"description": "Updates property with a valid id"
				,"tags": [
					"Update"
				]
				,"responses": {
					"200": {
						"description": "updated"
					}
					,"400": {
						"description": "bad request"
					}
					,"404": {
						"description": "property not found"
					}
				}
			}
		}
	}
	,"definitions":{
		"Property": {
					"properties": {
						"id": {
							"type": "integer"
						},
						"address": {
							"type": "string"
							,"minLength":1
							,"maxLength":255
						},
						"city": {
							"type": "string"
							,"minLength":1
							,"maxLength":50
						},
						"state": {
							"type": "string"
							,"minLength":2
							,"maxLength":2
						},
						"zip": {
							"type": "string"
							,"minLength":5
							,"maxLength":10
						}
					},
					"required": [
						"address",
						"city",
						"state",
						"zip"
					]
		}
	}
	,"responses":{ 
	},"parameters":{ 
	},"tags": [
		{ "name":"Get"
			,"description":"Get all properties" 
		},
		{ "name":"Get by ID"
			,"description":"Gets a selected property by id" 
		},
		{ "name":"Post"
			,"description":"Adds a property" 
		},
		{ "name":"hello"
			,"description":"Says hello to the caller" 
		},
		{ "name":"Delete by ID"
			,"description":"Deletes property selected by "
		},
		{ "name":"Update"
			,"description": "Updates property selected by id"
		}
	]
}